[Fix] #79 - DateFormat 기능 extension으로 전환
